requirements.txt
Django==2.2.10
pytz==2020.5
asgiref==3.3.1
sqlparse==0.4.1
djangorestframework
